Refactor prop types to its own file
